requirements.txt
apache-airflow==2.10.5
httpx==0.28.1
pandas==2.1.4
pendulum==3.0.0
SQLAlchemy==1.4.54
xlrd==2.0.1
openpyxl==3.1.5